refactor: move routes to files instead of folders chore: make vscode nest better
