+++ src/pages/$.tsx
@@ -1,0 +1,3 @@
+export default function MyNameIsError404() {
+  return <div>404</div>;
+}
