+++ src/pages/forum.thread.$threadId.tsx
@@ -1,5 +1,5 @@
 import type React from "react";
-import ForumThread from "../../components/forum/boards/thread.component";
+import ForumThread from "../components/forum/boards/thread.component";
 import { useParams } from "react-router";
 
 const ForumThreadPage: React.FC = () => {
