+++ src/pages/sonic._index.tsx
@@ -1,5 +1,5 @@
 import type React from "react";
-import Widget from "../../components/common/widgets/widget.component";
+import Widget from "../components/common/widgets/widget.component";
 
 const Sonic: React.FC = () => {
   return (
