+++ src/pages/user.profile.$userId.tsx
@@ -1,5 +1,5 @@
 import type React from "react";
-import UserProfileMaster from "../../components/user/userProfileMaster.component";
+import UserProfileMaster from "../components/user/userProfileMaster.component";
 import { useParams } from "react-router";
 
 const UserProfile: React.FC = () => {
